New implementation of the memory benchmark - configuring application to print detailed GC info and analyzing the result. JOPA and partially AliBaba.

+++ alibaba-benchmark/src/main/java/cz/cvut/kbss/benchmark/alibaba/data/AlibabaDataGenerator.java
@@ -5,7 +5,7 @@ import cz.cvut.kbss.benchmark.data.DataGenerator;
 
 public class AlibabaDataGenerator extends DataGenerator<Person, OccurrenceReport> {
 
-    public AlibabaDataGenerator(int factor) {
+    public AlibabaDataGenerator(float factor) {
         super(factor);
         generate();
     }
